Pass and utilise extension names in child process.
This change adds ability for the child process runner to get extension names from the parent process. This is necessary so that we can register a list of "supported" file extensions and later so that node can automatically resolve filenames for us. E.g. when there is a file named foo.jsx and node doesn't have a loader for such an extension it would try to append '.js' and load the file with '.js' as an extension. If '.jsx' loader is registered instead, node will try to use it.